feat: Align module with Snowflake provider changes
Due to breaking changes in Snowflake provider and additional code optimizations, **breaking changes** were introduced in this module.

List of code and variable (API) changes:
- snowflake_integration_grant resource has been removed and replaced by getindata/role/snowflake module
- minimum Snowflake provider version is `0.94.0`

For more information, refer to [variables.tf](variables.tf), list of inputs below and Snowflake provider documentation

When upgrading from `v1.x`, expect most of the resources to be recreated - if recreation is impossible, then it is possible to import some existing resources.